Add io.ReadSeeker compatibility (partially)
Change most ocurrences of os.File to io.ReadSeeker and create different
functions to handle from filenames. Unfortunately the Parse func still
can't receive a io.ReadSeeker as the original code requires the size of
the byte slice to be known from the beginning, and the io.ReadSeeker
does not provide such functionality.

Meanwhile I decided to use bytes.Reader instead, so hopefully if someone
only has the very first bytes of a mp4 file, they can still use the
Parse func.

I tried not to change in any way the error handling as I don't know if
handling an error may cause an unexpected behavior. However, there were
several file opens that were not closed, causing a potential resource
leak. The errors still are not handled in a way I am satisfied, but so
far the parsing has been working locally for me.

I think it needs a review or two for this to be production-safe.