Improve error handling in the mp4 package
Return error instead of panicking when errors rise, so the one importing
this package can decide how to handle the error.

While doing this, I discovered there is a risk associated with using the
parser as it will not parse blocks from unknown versions. This may cause
unexpected behaviour when parsing files with new -hence unknown- verions.

Also, changed MdatBox.Bytes() to a no-op func as it won't work anymore
with the io.ReadSeeker compatibility. This is becasue it uses os.Open
with mdat.Filename (mp4.go:2235), which was set when the Parse func
received a *os.File as a parameter. Due to the filename being an empty
string, it would have always failed.

All these changes are to easen the use of the mp4 parser as a library
in other go codebases. Thats why I am choosing to delegate most of the
error handling to whoever uses the library.