Add length validation for the avcc and sdtp boxes
This commit aims to avoid panics when the data contained in the box does
not follow the 'contract' 'specified' in the code, for the avcc and sdtp
boxes. This means the entirety of the data of these boxes wont be parsed
in those instances.

Following the original "error handling" of this codebase, I added debug
logs when such cases happen. I do this to avoid thinking into a better
error handling and having to rewrite all with that new error handling.

This is clearly dangerous as it will fail silently when the data cannot
be read, but as I'm not using that parsed data, I selfishly leave it
that way.