Skip read box func when (boxSize - 8) is 0
This change is introduced due to a panic caused by a file that has the
avcC block with size 8. I was afraid it could cause the reader to get
lost in its read offset but it turns out it really does not affect that
much. I tested against a set of 85 videos with various codecs and it
didn't blow up.